Refactor code to use PostRepository for database operations
Introduced a custom repository for Posts, `PostRepository`, and injected it into the `Post` entity. This allows us to perform database operations related to Posts in a more organized and isolated manner, enhancing code readability and maintainability. In addition, moved the database setup steps (Entity Manager & Post Repository) to `setupDatabase()`, updated PostResolver to use `PostService`, and renamed main function to `bootstrap`.