Implement createPost mutation in GraphQL resolver
The commit introduces a new 'createPost' mutation. The imported 'Mutation' from 'type-graphql' is used to define the resolver method for creating a new post. The 'title' argument is passed to this method. The PostService class has a new 'create' method to persist a new post data. The 'EntityManager' from 'mikro-orm/core' is used to interact with the database. This change will allow consumers of the GraphQL API to create new posts by providing a title.