Add updatePost and deletePost mutations to GraphQL resolver
The commit adds additional mutation methods 'updatePost' and 'deletePost' to the Post resolver in order to handle post modification and deletion. For each method, a corresponding logic has been added in PostService to interact with a database. 'EntityManager' and 'wrap' from 'mikro-orm/core' are used to manage persistence layer. This development greatly extends the flexibility and functionality of our GraphQL API regarding post data manipulation.