Add User table and update migration snapshot file
A new "User" table has been added in the Migration20231029185957.ts file with necessary attributes such as "id", "created_at", "updated_at", "username", and "password". Also included is a unique constraint on the "username" field. The mikro-orm.config.ts file has been updated to include the new User entity. Additionally, the migration snapshot-breddit.json file was modified to reflect this change. This is part of a larger effort to restructure the application database and accommodate new user-related features.