Add UserResolver and enhance session handling
This commit introduces the UserResolver for authenticated users and renames SignInResponse and SignUpResponse classes to SignInResult and SignUpResult respectively for better semantic understanding. We have also made changes to the session handling. The session cookie duration has been increased to 7 days and the session id has been renamed to 'breddid_session_id' to align it with our application. Storing of user's ID in the session post successful sign up has also been handled to enhance the user login process. Also, we have added a method in UserService to retrieve user by ID which will be helpful in various use-cases. The UserResolver is also imported in the apollo-server setup to cater GraphQL queries related to user data.