Add AuthorizationError class in GraphQL error types
A new error class, AuthorizationError, has been added to GraphQL error types. This tailored error class is then utilized in user resolver's methods to better deal with cases when the user is not authorized. The inclusion of this error type enhances the specificity and interpretability of our error handling mechanisms.